refactor: NumbersGenerator, RandomNumbersGenerator,  LottoStore lotto 패키지 이동
