+++ src/main/java/lotto/domain/lotto/NumbersGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain.draw;
+package lotto.domain.lotto;
 
 import java.util.List;
