+++ src/main/java/lotto/domain/lotto/RandomNumbersGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain.draw;
+package lotto.domain.lotto;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
