+++ src/main/java/lotto/service/LottoService.java
@@ -3,8 +3,8 @@ package lotto.service;
 import java.util.List;
 import lotto.domain.draw.DrawingMachine;
 import lotto.domain.draw.DrawingResult;
-import lotto.domain.draw.LottoStore;
 import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoStore;
 import lotto.domain.lotto.Lottos;
 import lotto.domain.lotto.WinningCombination;
 import lotto.domain.money.Money;
